style(core): unify LogLevel enum naming to UPPER_CASE convention
- Rename LogLevel enumerators: Trace→TRACE, Debug→DEBUG, Info→INFO,
  Warn→WARN, Error→ERROR, Fatal→FATAL, Off→OFF
- Update all usages: logger macros, tests, examples, ROS2 integration
- Enforce brace-style for single-line if statements
- Remove unused headers (e.g., <thread> from log_context.cpp)

Files: 29 files across core, tests, ros2, and examples

+++ br_logger_ros2/include/br_logger_ros2/ros2_bridge.hpp
@@ -15,7 +15,7 @@ struct BridgeConfig
   bool enable_console = true;
   bool enable_file = false;
   std::string file_path = "/tmp/robot_logs/";
-  size_t max_file_size = 50 * 1024 * 1024;
+  size_t max_file_size = static_cast<size_t>(50 * 1024 * 1024);
   uint32_t max_files = 5;
   std::string console_pattern = "[%D %T%e] [%C%L%R] [%g] [%f:%#::%n] %m";
 };
